Improved initial package handling
Before blindly yum installing the packages needed by these scripts in a
few spots, check if the packages are already installed with just the rpm
command. This can end up saving a decent amount of time when iterating
on running these scripts locally.

Also adds an install of wget, since that package is not always installed
on CentOS systems.

Change-Id: Ic3bc315337e2237c348d9cd0d6eb499bc66c7d9e