Pass $MTU env var to toci_vxlan_networking.sh
The $MTU var needs to be passed to the undercloud-setup role so that
when the toci_vxlan_networking.sh script is executed the value is
honored.

Often, a lower value needs to be set than is defaulted in tripleo.sh
(1450). It depends on how the host cloud is configured (which isn't
always known without experimentation).

Change-Id: Ife5225cfb009f7e170a3d71ca8fc9aae0725caf7