ovb/multinode: cleanup rdo-release/puppet packages
OpenStack Infra provides images with rdo-release and puppet packages,
which is something we don't want in TripleO CI because we already have
our own packages from local mirror, depending of the version of
OpenStack that we want to test.

This patch makes consistent the cleanup on both OVB & multinode jobs to
remove puppet & rdo-release packages.

Closes-Bug: #1670015
Change-Id: I5f6b6be6f07dfd7d27189547db9daf1594db8c85