Run overcloud-upgrade role only in upgrades jobs.
Defined tags are the same for all multinode
jobs, independently on if they perform an
upgrade or not. Each featureset env file then
defines if an upgrade is to be performed or not,
being skipped all the tasks of the overcloud-upgrade
role if it's not an upgrades job.

This task skipping takes around 3 seconds,
what's not a big deal, but it makes the logs
more difficult to read and understand.

This patch appends the overcloud-upgrade role
into the tags only if the job to be run performs
an upgrade.

Change-Id: Ia037dd74470499d302cb098136fbe71608b7824a