Revert "Install ssh keys of tripleo ci admins on nodes"
It was highly inappropriate to add new ci admins in this change,
especially without mentioning it in the commit message.
Note that strictly speaking we require a quorum of the
existing admins to approve adding new ones, although with the
number of admins growing we may want to revisit that policy.
As of this time it is still in effect, however.[1]

And even if it weren't, it shouldn't have been
piggy-backed on this change.  Please resubmit as two separate
changes so the admin change can be properly reviewed.

1: Unfortunately we lost the documentation of this when the
admin list was moved from tripleo-incubator to tripleo-ci:
https://github.com/openstack/tripleo-incubator/blob/4de13b35d91150eb269bea22820d0e688cd6b643/tripleo-cloud/README.md
That should be fixed too.

This reverts commit a4d1f8c9605299053ec3e4b19f4b885ea2d0619f.

Change-Id: I6a5484dfb1c00431ae89ec20d9dd66b623b3e622