zuul: Run OVB jobs only when it matters
There is no need to execute OVB jobs when we only patch files that
deploy services not deployed by OVB (but by scenarios), so let's save
resources and exclude files like featuresets, manifests or tripleo
services that are not used in any of our OVB jobs.

It introduces a child job, so we can still use the same irrelevant-files
as multinode jobs.

Change-Id: I8216328b40a8a430b664786a0b0896be8d1f16ff