Enforce a minimum TTL for DNS records in TripleO jobs
We have reasons to believe there is a correlation between DNS lookup
failures we are seeing in the gate and the very low TTL of some
domains.

Since jobs are running on ephemeral nodes which wouldn't be impacted
by stale cache, enforce a minimum TTL of 900 (15 minutes) for DNS
records in the local Unbound configuration.
This means that a domain with a TTL of 60 seconds will be cached for
15 minutes instead.

This will lead to less queries which is less prone to failure and rate
limiting by recursive DNS providers.

Change-Id: Iab9f4426304fab6b9e6f0920ad5c4e3b94d74954