Use ansible-playbook in place of quickstart.sh
This will change CI to only use quicktart.sh to bootstrap the
virtualenv. The ansible-playbook command will then be run
directly for the deploy and collect logs.

This is a first step in breaking the deploy into multiple
ansible-playbook calls so that we CI the ability to run them in
stages for development purposes.

Change-Id: Ie0a3729277fb608c653e7bc2ab85781d9b815880