Remove hardcoding of overcloud deploy timeout
Now overcloud deploy timeout is derived from DEVSTACK_GATE_TIMEOUT
which is always 170 and equal to 80. It's not appropriate for all
jobs. It will be calculated dynamically depending on how much time
remains in the job for deployment.
Pass end time of CI job to ansible playbook for calculating it.

Partial-Bug: #1738038
Change-Id: I89b743fc16ea4c100ba21ddbe29081fa3e5479e9